Created basic infrastructure & contain expectation
This commit creates the basic structures to be used by cppshould, and
an initial draft of the Contain matcher.

Details of the basic types:

ExpectationTraits - A template that should be specialized for each
                    expectaion (and potentially data-type) to define
                    the behaviour of the expectation.
BaseExpectation -   The base class for all expectations.
Expectation -       The base expectation template, depends on the
                    type being passed in to the expectation, and
                    should define type-specific behaviour
ShouldInfo -        Takes information from the SHOULD macro and
                    stores it till needed by the Should class
Should -            The main Should class responsible for checking
                    the expectations (using ExpectationTraits)