Added SConscript to build google test framework
The latest version of ubuntu does not provide a static library for the
google test framework, as building the framework with different compiler
flags from your actual application can cause odd errors.

This commit adds in a SConscript to the deps folder that will build
gtest from source if it is installed in the /usr/src folder as it is on
ubuntu.  This should allow travis-ci to build the unit tests.

In the future I might update things to not neccesarily rely on the
source being in this location, but for now this should do the trick.