Updated gtest build code to download the source
The last commit caused scons to always attempt to build a local copy of
gtest even if there was a copy installed globally.  Of course this
didn't work if there wasn't a copy of the source code in /usr/src/gtest.

This commit updates the gtest build code to look for a copy in
/usr/src/gtest but if it's not found, then it will download a copy from
google code and then compile that instead.