Removed option of using gtest from /usr/src
When using a copy of the gtest source from /usr/src scons was trying to
compile to /usr/src which obviously the user does not have access to.

To get around this difficulty, I've just decided to always make use of
the in-tree version of the source code, and remove the option of using
/usr/src.

I've also removed the install of libgtest-dev from .travis.yml as
there's no longer any point.