Switched to unzip instead of tar for gtest
The version of tar installed on linux doesn't support zip files (whereas
the mac version i tested with does) so this changset updates to use
unzip instead.

I've also discovered a minor bug where the source won't be downloaded if
the build folders already exist.  Not sure why this is, but it seems
os.path.exists is using the build folder.  I might fix this in the
future, but for now I can't be bothered.