Moved ExpectationTraits & added EquivalenceTraits
This commit moves ExpectationTraits out into the top-level traits file,
and creates EquivalenceTraits.

From a documentation point of view it makes a bit more sense to have
these in the one location as they're all used for cusomising behaviour.
Also added some limited documentation to this effect, although it'll
need to be updated substantially later.