Created ToStringTraits for converting to strings
Previously was just using ostream operator<< to convert things to
strings, but this was causing problems with outputting any
std::vectors (couldn't find the operator<< as it wasn't defined before
it was used and argument dependant lookup only finds things in the same
namespace as arguments).

This commit creates a new traits class to handle string conversion and
implements a specialization of it for std::vector.  Other stdlib
containers will be done at a later date.