Refactored building of expectation errors
Instead of having 2 functions that are implemented by each expectation
this commit updates things to have a single function that is defined in
the base class and called with arguments.

That function can be overridden by child classes, but provides some sane
default behaviour that delegates to a "ToString" class that should
return a string describing the expectation.  This way most of the work
can be done by the base class.

The base class now also differentiates between should & must